Разбил utils на отдельные модули
