+++ src/utils/events.js
@@ -1,0 +1,1 @@
+export const isEscEvent = (evt) => evt.key === 'Escape' || evt.key === 'Esc';
