+++ src/view/films-list.js
@@ -1,4 +1,4 @@
-import { createElement } from '../utils.js';
+import { createElement } from '../utils';
 
 const createFilmsListTemplate = () => (
   `<section class="films">
